chore: opt codes
