+++ swanlab/data/__init__.py
@@ -15,7 +15,6 @@ from .run import (
     get_config,
     get_url,
     get_project_url,
-    get_run_dir,
     config,
 )
 from .sdk import (
