+++ swanlab/log/__init__.py
@@ -6,9 +6,8 @@ r"""
 @IDE: vscode
 @Description:
     日志记录模块，在设计上swanlog作为一个独立的模块被使用
-    FIXME: shit code
 """
-from .log import SwanLog, trace_handler
+from .log import SwanLog
 
 swanlog: SwanLog = SwanLog("swanlab")
 
@@ -16,5 +15,4 @@ start_proxy = swanlog.start_proxy
 
 reset = swanlog.reset
 
-__all__ = ["start_proxy", "reset", "swanlog", "trace_handler"]
-
+__all__ = ["start_proxy", "reset", "swanlog"]
